+++ packages/apis/lib/network/remote/orders/abandoned_checkouts/api/api_abandoned_checkouts_service.dart
@@ -1,6 +1,6 @@
 import 'package:apis/apis.dart';
 import 'package:apis/dio_config/api_dio_client.dart';
-import 'package:apis/network/remote/orders/abandoned_checkouts/abstract/abandoned_checkouts.dart';
+import 'package:apis/network/remote/orders/abandoned_checkouts/abstract/abandoned_checkouts_service.dart';
 import 'package:apis/network/remote/orders/abandoned_checkouts/freezed_model/response/abandoned_checkouts_count_response.dart';
 import 'package:apis/network/remote/orders/abandoned_checkouts/freezed_model/response/abandoned_checkouts_list_response.dart';
 import 'package:dio/dio.dart';
